Feature/api endpoints (#2)
* Added: user routes

All the routes related to the user were added.

* Fixed: User routes

The logs were updated. Was added an route /user/login so the user can be logged without his id.

* Fixed: Swagger documentation

Were fixed some entries in the swagger documentation

* Added: Watchlist routes

Were added the routes for controlling watchlists. Also, were added to the models a Request and a Response Struct becoming more convenient to manipulate the req/res.

* Added: Media routes

The media routes were added. The swagger documentation was updated

* Fixed: API errors

Some errors were fixed in the logic behind the models crud